--- src/contexts/DpiHistoricalMarketData/index.ts
@@ -1,2 +1,0 @@
-export { default as DpiMarketDataContext } from './DpiMarketDataContext'
-export { default as DpiMarketDataProvider } from './DpiMarketDataProvider'
