+++ src/contexts/DpiTokenMarketData/DpiTokenMarketDataContext.ts
@@ -1,0 +1,14 @@
+import { createContext } from 'react'
+
+interface DpiTokenMarketDataValues {
+  prices?: number[][]
+  marketcaps?: number[][]
+  volumes?: number[][]
+  latestPrice?: number
+  latestMarketCap?: number
+  latestVolume?: number
+}
+
+const DpiMarketData = createContext<DpiTokenMarketDataValues>({})
+
+export default DpiMarketData
