+++ src/contexts/DpiTokenMarketData/index.ts
@@ -1,0 +1,2 @@
+export { default as DpiTokenMarketDataContext } from './DpiTokenMarketDataContext'
+export { default as DpiTokenMarketDataProvider } from './DpiTokenMarketDataProvider'
