--- src/hooks/useDpiMarketData.ts
@@ -1,9 +1,0 @@
-import { useContext } from 'react'
-
-import { DpiMarketDataContext } from 'contexts/DpiHistoricalMarketData'
-
-const useMarketData = () => {
-  return { ...useContext(DpiMarketDataContext) }
-}
-
-export default useMarketData
