+++ src/hooks/useDpiTokenMarketData.ts
@@ -1,0 +1,9 @@
+import { useContext } from 'react'
+
+import { DpiTokenMarketDataContext } from 'contexts/DpiTokenMarketData'
+
+const useDpiTokenMarketData = () => {
+  return { ...useContext(DpiTokenMarketDataContext) }
+}
+
+export default useDpiTokenMarketData
