+++ src/utils/coingeckoApi.ts
@@ -1,4 +1,3 @@
-
 const baseURL = 'https://api.coingecko.com/api/v3'
 
 export const fetchHistoricalTokenMarketData = (
